+++ config.php
@@ -4,8 +4,8 @@ use humhub\modules\dashboard\widgets\Sidebar as DashboardSidebar;
 use humhub\modules\space\widgets\HeaderControlsMenu;
 use humhub\modules\space\widgets\Menu;
 use humhub\modules\space\widgets\Sidebar;
-use humhub\modules\user\widgets\AccountTopMenu;
 use humhub\modules\user\widgets\ProfileMenu;
+use humhub\modules\admin\widgets\AdminMenu;
 use humhub\widgets\TopMenu;
 use k7zz\humhub\bbb\Events;
 
@@ -20,7 +20,7 @@ return [
         ['class' => Menu::class, 'event' => Menu::EVENT_INIT, 'callback' => [Events::class, 'onSpaceMenuInit']],
         ['class' => ProfileMenu::class, 'event' => ProfileMenu::EVENT_INIT, 'callback' => [Events::class, 'onProfileMenuInit']],
         ['class' => HeaderControlsMenu::class, 'event' => HeaderControlsMenu::EVENT_INIT, 'callback' => [Events::class, 'onSpaceHeaderControlsMenuInit']],
-        ['class' => AccountTopMenu::class, 'event' => AccountTopMenu::EVENT_INIT, 'callback' => [Events::class, 'onAccountTopMenuInit']],
+        ['class' => AdminMenu::class, 'event' => AdminMenu::EVENT_INIT, 'callback' => [Events::class, 'onAdminMenuInit']],
         ['class' => Sidebar::class, 'event' => Sidebar::EVENT_INIT, 'callback' => [Events::class, 'onSpaceSidebarInit']],
         ['class' => DashboardSidebar::class, 'event' => DashboardSidebar::EVENT_INIT, 'callback' => [Events::class, 'onDashboardSidebarInit']],
     ],